requirements.txt
Django~=5.2
djangorestframework~=3.16.0
djangorestframework-simplejwt
psycopg2-binary
pytest~=8.3.5